# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.SureSureUser:
    properties:
      access_token:
        type: string
      address:
        type: string
      bill_date:
        type: string
      created_date:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: integer
      merchant_id:
        type: integer
      name_en:
        type: string
      name_th:
        type: string
      package_change_date:
        type: string
      package_id:
        type: integer
      password:
        type: string
      phone:
        type: string
      picture:
        type: string
      quota_all:
        type: integer
      quota_left:
        type: integer
      quota_usage:
        type: integer
      step:
        type: integer
      store_category_type:
        type: string
      store_email:
        type: string
      store_name:
        type: string
      store_phone:
        type: string
      token:
        type: string
      uid:
        type: string
      updated_date:
        type: string
      user_role:
        type: string
      user_type:
        type: string
      username:
        type: string
      website:
        type: string
    type: object
  util.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      timestamp:
        type: string
    type: object
info:
  contact: {}
  description: Public API for SureSure services
  title: SureSure Public API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Create user if not exists, otherwise return existing user
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.SureSureUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Login or create user
      tags:
      - User
...
  /user/category/get:
    get:
      description: Retrieve merchant categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: List categories
      tags:
      - User
  /user/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.SureSureUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Create user
      tags:
      - User
  /user/delete/{id}:
    delete:
      parameters:
      - description: User UID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Delete user by UID
      tags:
      - User
  /user/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: List users
      tags:
      - User
  /user/get/{id}:
    get:
      description: Retrieve a user record by UID
      parameters:
      - description: User UID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Get user by UID
      tags:
      - User
  /user/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.SureSureUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Update user
      tags:
      - User
schemes:
- http
- https
swagger: "2.0"